Add pcodeops for setting SPRs and reading TB
Gamecube system code often lacks a lot of necessary operations to
do with setting SPR values.  This adds pcodeops to force the
decompiler to show these SPR writes as "setSpr(SPR, X)".

Additionally, looping on a TBR read could sometimes decompile
as an infinite loop, causing later parts of the function to not
show at all in the decompiler window.

This is intended as an optional variation on the Gekko processor
definition, rather than the default one to be used by everyone.